Refactor: Move useContent smoke test to useContent.spec.jsx
Closes #210
